+++ react-router.config.ts
@@ -1,6 +1,6 @@
 import type { Config } from "@react-router/dev/config";
 
 export default {
-  basename: process.env.GITHUB_ACTIONS ? "/personal-ai-portfolio" : "/",
+  basename: process.env.GITHUB_ACTIONS ? "/liangshanbobo" : "/",
   ssr: true,
 } satisfies Config;