+++ vite.config.ts
@@ -3,7 +3,7 @@ import tailwindcss from "@tailwindcss/vite";
 import { defineConfig } from "vite";
 
 export default defineConfig({
-  base: process.env.GITHUB_ACTIONS ? "/personal-ai-portfolio/" : "/",
+  base: process.env.GITHUB_ACTIONS ? "/liangshanbobo/" : "/",
   plugins: [tailwindcss(), reactRouter()],
   resolve: {
     tsconfigPaths: true,